demo/requirements.txt
asgiref==3.4.1
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==3.2.9
django-appconf==1.0.5
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==4.0.1
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
mysqlclient==2.1.0
packaging==21.3
pdfkit==1.0.0
pilkit==2.0
Pillow==8.4.0
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.12.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
